Access to info about feature constraints refactored
