+++ FlintCore/Constraints/Permissions/SystemPermissionRequestAction.swift
@@ -1,0 +1,15 @@
+//
+//  SystemPermissionRequestAction.swift
+//  FlintCore
+//
+//  Created by Marc Palmer on 09/05/2018.
+//  Copyright © 2018 Montana Floss Co. Ltd. All rights reserved.
+//
+
+import Foundation
+
+public enum SystemPermissionRequestAction {
+    case requestPermission
+    case skipPermission
+    case cancelAll
+}
